UHF-11974: Dividing CSP directives into multiple context specific event subscribers.
