+++ helfi_platform_config.services.yml
@@ -24,8 +24,32 @@ services:
     class: Drupal\helfi_platform_config\EventSubscriber\SearchApiSubscriber
     arguments: []
 
-  Drupal\helfi_platform_config\EventSubscriber\CspEventSubscriber:
-    class: Drupal\helfi_platform_config\EventSubscriber\CspEventSubscriber
+  Drupal\helfi_platform_config\EventSubscriber\CspCleanSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspCleanSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspIbmChatAppSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspIbmChatAppSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspReactAndShareSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspReactAndShareSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspTeliaAceWidgetSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspTeliaAceWidgetSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspElasticProxySubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspElasticProxySubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspCommonSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspCommonSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspLocalDevSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspLocalDevSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspSiteimproveSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspSiteimproveSubscriber
+
+  Drupal\helfi_platform_config\EventSubscriber\CspSocialMediaSubscriber:
+    class: Drupal\helfi_platform_config\EventSubscriber\CspSocialMediaSubscriber
 
   Drupal\helfi_platform_config\EntityVersionMatcher: '@helfi_platform_config.entity_version_matcher'
   helfi_platform_config.entity_version_matcher:
